[mysqli] Stop tracing mysqli_fetch_field_* methods
We were warned by users that the mysqli integration was generating a large number of
spans related to the fetch_field_* methods. Those methods are executed many times per query
(one for each fetched field) and can result in 80%+ of the spans generated per single
mysqli query.

Moreover info in such spans are not very interesting, as the real information there
is in the amount of time the query took to be execute and the amount of time
results took to be loaded in memory.

For this very reason, we removed tracing of 'fetch_field_*'.