[mysqli] Do not trace fecth methods
There is not real value in tracing mysqli's fetch methods as when invoking
specific methods, e.g. 'mysqli_fetch_row' the row is not actually retrieved
from the database, but it is simply read from the socket stream and translate
from the C extension domain to the PHP domain as a PHP array.
If the driver would have gone to the database, then some interesting info,
e.g. network latency or errors, may have been relevant, but given this consideration
above we decided to drop them entirely and implement, on a different PR, a more
sofisticated analysis, e.g. only reporteing slow fetch operations based on
p99 analysis ir similar.
Implementation yet to be decided.