[laravel] Add Laravel 5.8 compatibility
Prior to Laravel 5.8 'events' was available in container during first app()->bind()
because aliases were already registered.

With Laravel 5.8 this behavior changed, resulting in an error due to missing 'events' alias
in container.

Based on our latest changes that allow us to trace classes and methods that no longer exists,
we refactored the part of code which binds to events to directly trace event's constructor methods.