Put source files in a sorted multi-line variable
Prior to this, whenever the sources list changed it would be
difficult to review. This should make it much easier.

Note that convention in many other projects puts the "main" source
file in the first of the list. Since the "ddtrace.so" file is made
from "ddtrace.c" I put that one first.