avoid `X-Datadog-Trace-Count` error in agent during dd-doctor.php requests
When we ask customers to enable dd-doctor.php, then we find in the agent
logs the following error which is misleading:

```
2019-12-23 10:00:16 PST | TRACE | WARN | (pkg/trace/api/api.go:358 in handleTraces) | Error getting trace count: "HTTP header \"X-Datadog-Trace-Count\" not found". Functionality may be limited.
```

The reason for that is that dd-doctor.php sends an empty array of traces to
the agent to check connectivity. This request this not provide the header above
(it is created manually).

This PR adds the above header to all requests with a value of `X-Datadog-Trace-Count: 0`.