make sure traces are flushed when in limited mode
Release 0.37.0 introduced to changes that, together, prevent traces to be sent when the span limit is reached.

Specifically:
 - #689 prevent a callback registered through dd_trace_(method|function) to be called if we are in limited mode,
   AKA if the number of spans exceed the value DD_TRACE_SPANS_LIMIT which defaults to 1000.
 - #707 sandboxed the execution of the final flush of the tracer using, as a hook, a callback registered through
   dd_trace_method().

As a consequence, when the limit of DD_TRACE_SPANS_LIMIT was reached (default is 1000) no more callbacks are
invoked anymore, including the final one that would trigger sending traces to the agent.