Cleanup finding dispatch and add cputime.d dtrace script
The cputime.d dtrace script showed ddtrace_find_dispatch as a
function of interest for performance.  This commit does not
change peformance very much; basically there is a addref/delref
that is avoided. However, I think the code is better and the
cputime.d script will probably be useful later.