Make minor robustness tweaks to the tests
After testing on the nice-neighbor framework, there were a number of false positives resulting from sending output from the request init hook. Some extensions modify the header in their RINIT and this would cause a, "Headers already sent" E_WARNING.

This also includes a tweak to the memory-limit test to ensure that it will still cause a fatal error when there is a real request init hook set from the INI settings.