Avoid sigsegv on test infrastructure
If the builtin webserver does not get output then it will sometimes
sigsegv. We've hit this case many times. I think the only way we
can stop hitting this case is to not use the built-in CLI webserver.

At the same time, there are constant propagation bits in the compiler
that seem to influence this particular case; if I use fully-qualified
names then even without generating output the issue goes away.