Guard span_fci
If the tracer flushes during something like trace_method then the
internal span stack it might go away. The tracer currently uses a
tracing posthook to flush, which is how this was found; it should
be migrated to a non-tracing hook but we want to make sure this
case doesn't segfault if it comes up again.