add composer.json file for src/api tests
We want a separate composer.json file to test the src/api suite.
The reason for that is we do not want a dependency on it while we test the regular automatic
instrumentation scenario, where the src/api package is not entirely available.

NOTE: as you can see we depends on src/api via "autoload.psr-4". This is not optimal and only
an approximation of what the user scenario will be. Unfortunately we cannot use path repos here
because the folder is in the main package. Applying tricks (e.g. copying or archiving as symlink is not working) would
make things way to complex. We keep it like here until we will be able to have a separate (and indipendent) shim to import.