Call ddtrace_engine_hooks_r{init,shutdown}
These were added, but then never called :facepalm:

Honestly, we were unlucky that tests did not catch this. This means
that most of the time the function used the same address in subsequent
requests.

I also don't know how to add a test for this, as I don't know how to
force the address to change.