Makefile: do not optimize Laravel in tests
The reason is that during optimization the environment has to be 'setup'.
E.g. the config is dumped and `app.name: getenv('APP_NAME', 'Laravel')` becaomes `Laravel`
unless the env APP_NAME is not setup during optmization.

In our case envs are set while running the test, so APP_NAME is not available yet.