Make execution of ext tests locally faster and easier
As part of every execution, before we copied ALL files (sources, tests, .m4) and run a complete
build even if only one single file changed. With this PR we separate the concept of source files and test files.

When a test file changes, the file is only automatically copied to `tmp/build_extension/tests/ext/**` and a build is NOT done.
When a source file changes, the file is copied to `tmp/build_extension/**` and a build IS done automatically.

Currently developers can do `make test_c TESTS=tests/ext/some_thig.phpt` and the quickest possible process will be executed, only
building if required.

Note that the `EXTRA_CFLAG` variable is provided to provide extra CFLAGS, e.g. `EXTRA_CFLAGS=-g`.

Limitations:
  - When a `.h` file changes, it is automatically copied to thebuild directory but a build it NOT done.