[Tracer] Add runtime id, components, and plugins
Adds top-level directories components and plugins. Components have
a single public header and their own tests. Plugins are a type of
component that are coupled to the PHP lifecycle.

The runtime id is used to associate traces to other data. It is
unique per process and should be regenerated after a fork. The
profiler will use the runtime id to associate profiles to traces.

This adds a dependency on libsodium for generating random bytes.
I expect many CI jobs will fail because of this.

This bumps the minimum version of CMake required to 3.13. The
specific feature needed is:
 - The install(TARGETS) command learned to install targets created
   outside the current directory.

This is used to install all components from the common
components/CMakeLists.txt.