do not keep alive connections while executing randomized testing
A degree of flakiness in randomized testing was observed regardless of the tracer installed or not.

A typical error that is a false positive is [this one](https://app.circleci.com/pipelines/github/DataDog/dd-trace-php/5231/workflows/c839b1ba-0e9e-4f07-ab5e-9d304e121c01) where a EOF error is generated by vegeta supposedly when the server close the connection due to high traffic.

Searching online it seems that this is generated when the Go's http client receive a response with `Keep-Alive` and then the server closes the connection, for other reasons.

In order to increase stability we set to false the `-keepalive` flag in vegeta and observe flakiness in the next few pushes.