[Tracer] Add component system, string_view and sapi components
Adds top-level directory components. Components have a single
public header and their own tests.

This bumps the minimum version of CMake required to 3.14. The
specific feature needed is:

> The install(TARGETS) command learned how to install to an
> appropriate default directory for a given target type, based on
> variables from the GNUInstallDirs module and built-in defaults,
> in lieu of a DESTINATION argument.

This is so each component doesn't have to copy/paste the same config
and keep it in sync.

This also drops the older Catch2 version; it was not being done
for any specific reason.

Put bad test on PHP 5.4 into xfail; already xfail this on PHP 5.5.
php/php-src@471540d