Better safeguarding in dd_send
During a send-type operation, parameters are checked to see whether they
are required.  Before, if a parameter was required, an error was thrown
only when that parameter also had no default value (why would you have a
default value for a required paramater, you may ask--nearsightedness is
why)

This patch very simply overrides the default column to specify an error
to be given when the value is absent.  It also slightly restructures the
order of defaulting in the handler that populates the HTTP header.