Begin documenting architecture
This is an incomplete documentation of components, ext/php$n, and
the background sender. This information should be valuable to
newcomers.

I opted to write this as markdown in this repository instead of
a Google Document so that it is close to the thing it documents
and is open to everyone, not just Datadog employees.