Laravel for PHP 7.3 and 8.0 in rel env with static pool configuration
We configured the main container in the rel env to have 4 cpu cores.
We need to keep the number of process low for Laravel since the runtime load is pretty much
CPU-only (it is a synthetic app that does zero calls to db etc), so we would immediately saturate the CPU
and this would make calculating the CPU overhead impossible.