Add arena component
This arena is used to do bump pointer style allocations inside of a
previously allocated block, such as stack memory, static storage, or
heap memory.

Individual allocations made by the arena are not tracked and are not
freed. The use-case is allocating a bunch of data which has the same
lifetime, and then everything can be freed at once by resetting the
arena. This means that care must be taken to either not store types
which have custom destructors, or that they should be tracked and dtor'd
outside of the arena before resetting it.