Add uuid component
Represents a UUID abstractly, but mostly for the v4 variant which uses
random bytes.

This will be used in the Code Hotspots feature. The `runtime-id` is a
UUIDv4 that is shared between traces and profiles in the same process.
Code Hotspots requires more information than just the runtime id, but
this is one piece to it.