Zai/json updates (#1397)
This fixes a few different kinds of CI failure we were experiencing since adding ZAI/json. 

In verify builds we appeared to have a problem caused by symbol clashes, fixed by not declaring symbols with the same name as symbols that might be statically linked into php.

We had a spurious failure caused by non null-terminated strings, fixed that in tests and commented in header.