Fix randomized testsuite memory limits and background sender curl handling (#1525)
* Increase accepted average memory limit in randomized long running tests

With 5MB limit we were at the very limit. Whit the recent changes to ES integration
we added a bit of memory overhead and we went right above in certain cases. For those same
tests we were, for example, 4.94MB avg before, 5.05 after, triggering the error.

Optimization of single integrations will have to be done but for now it is out of scope for the
work on ES.

* Disable signal handling as recommended per libcurl threadsafety docs

See: https://curl.se/libcurl/c/threadsafe.html

Co-authored-by: Luca Abbati <luca.abbati@datadoghq.com>