Remove regressions from randomized tests
Since the very beginning, it never happened that a frozen regression test in the randomized tests caught an error
that the 'regular' randomly generated scenarios would not.

Moreover, maintaining frozen regressions is harder because the suite evolves and it forces us to consider each (or a batch of) regressions
as a special case.

For these reasons we decided to remove them.