Fix dispatch resetting
On PHP 8 we cache our dispatches directly with the op_arrays. We thus must not free the dispatches before we are certain that we won't use them anymore. As an additional safeguard the tracer disables itself in RSHUTDOWN.

Given that on PHP 5 we also preserve dispatches across tracer disabling, we bring this behaviour in line for all versions and just keep the dispatches when disabling (or re-enabling) the tracer.

Fixes a crash reported in #1545.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>