Remove Symfony 4.2 from PHP 7.4 web tests
Symfony 4.2 did not resolve to a properly installable set of packages when using symfony/flex, which was fixed with https://github.com/symfony/flex/commit/e74319d17ed4edc5f27b9d24757e172930a7ebc4.
This incidentally made the composer run as a whole not fail. Now that this is working now, a zend-code compatibility layer seems to be missing.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>