(PHP 7.3+) Fix compatibility with opcache (#1656)
In the initial commit of the interceptor, we intended to be executed after opcache, to iterate over the loaded classes and functions, but this was never properly tested (in part because the opcache tests were run with DD_TRACE_CLI_ENABLED not set).
It turns out that it did properly work on PHP 7.0 - 7.2, as the opcache logic was happening in its startup handler.
However on PHP 7.3 and 7.4, opcache was moved to post_startup, causing it to be after our code. Similarly on PHP 8, we did use post_startup, but by mistake we first called our own code, then opcaches handler, defeating the the whole point of that code.

This has been fixed, the post_startup logic of opcache is now executed before our own, allowing us to read the result of zend_compile_file.

Additionally, the JIT does not allow for opcode overloads. By ensuring that our code installing the user_opcodes is executed after that check (happening in opcaches post_startup as well), we bypass that check.
However, given that the JIT is interacting very poorly with EX(opline) replacements (It assumes there is no code installing custom oplines...) as it has installed some metadata after the end of the op_array at some offset we cannot access on PHP 8.0, we have no choice, but create custom specialized logic for the JIT.

Fixes #1654 and #1655.

* Enable JIT in randomized tests

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Try delaying resolver.c to post startup hook

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Actually run opcache tests with ddtrace enabled, add opcache JIT test

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Fix post_startup order vs opcache

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Use a special handler for intercepting function/class declarations with JIT active

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Add tests for interceptor function&class resolving with opcache enabled

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Fix the opcache issue on PHP 7.3 and 7.4 as well

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>