Adjust test so runtime-id is always available
This is simpler to implement in Rust because it avoid global mutable
state in favor of lazy initialization.

It does have the unfortunate side-effect that if the profiler is
installed but disabled and the user clicks on the Code Hotspots tab,
then the app will try to find associated profiles even though there
won't be any.