Build and package datadog-profiling in CI (#1663)
This adds jobs to the `build_packages` workflow in CircleCI that build
the datadog-profiling extension on x86_64-unknown-linux-gnu and
x86_64-alpine-linux-musl. It adjusts the `datadog-setup.php` install
script to use `extension = datadog-profiling.so` instead of
`zend_extension = datadog-profiling.so` and checks for profiling 
dependencies that weren't there before, which depending on your OS
includes `libgcc_s`, `libdl`,  `libpthread`, and `librt`.

This also updated the Alpine images to v3.16, which caused a few things
to break in CI. I fixed one of them already, but there's still an issue
with the warning `return-local-addr`. This was beyond my skill to just
drop in and fix quickly, so I dropped in a
`-Wno-error=return-local-addr` that should be removed once the fix is
in place. In the testing I did, the tracer built from this workflow can
run on Alpine 3.11 and earlier for x86-64, but I suspect for arm it
will not be possible to due to the libmusl v1.2 upgrade that happens
on Alpine 3.13.

* Switch from fat LTO to thin

This shaves off about 40s of link time. I haven't profiled performance
yet _at all_ so this is fine.