[profiling] Add Unix Domain Socket (UDS) support
If any configuration is set, it prioritizes that over UDS. If the
socket file doesn't exist then it falls back to http://localhost:8126.

There are tests for various units, and I have tested this manually for
end-to-end for both the default and setting `DD_TRACE_AGENT_URL` to
`unix:///var/run/datadog/apm.socket`.

Update to libdatadog v0.8.0.

Co-authored-by: Bob Weinand <bob.weinand@datadoghq.com>