Avoid some crashes with forking
This replaces the global profiler object in the child process with
None, and sets profiling_enabled to false. This leaks some memory
at the expense of crashing less.

However, it's still possible to have issues, for instance if a thread
in the process holds a lock at the time of the fork. So there is
further work to either synchronize such that no locks are held, or to
clean up all that stuff before the fork, and make it again after.