Mitigate panic in profiler shutdown, bound timeout
This mitigation is for issue #1919. It doesn't fix the root cause,
which we haven't been able to determine yet.

This also bounds our shutdown timeouts. Before, they _ought_ to have
been bounded as a consequence of how the code was written, but this
is explicit. It's also a tighter bound than before, 2 seconds instead
of 10 seconds.