perf(profiling): optimize prepare_sample_message
This function is called once per stalk walk, so it's on the hot path
but not as important to optimize as walking the stack itself.

However, it was showing up more than I would have expected in my
profiles when examining it with the native profiler. Unfortunately I
didn't have granular feedback on what was slowing it down, so I did
some general optimization techniques:

 1. Statically store the sample types as an array instead of creating
    them on-demand.
 2. Replace Cow<'static, str> with &'static str in ValueType. All
    usages were already static strings, and I don't see this changing.
 3. Simplify the branching logic into slicing operations.
 4. Move tag creation to an earlier point and do dumb-copies of tags
    when needed instead of branching logic.