fix(profiling): cleanup profiling status in a native fork
In case PHP-FPM is run with preloading as a non-root user, it will run
preloading in the `php-fpm: master process` which leads to our profiling
getting started there. When actual requests arrive, PHP-FPM will spawn a
child which then has the main thread and "thinks" profiling is running,
but all thread handles are invalid and pointing to the threads in the
master process.