feat(profiling) move upscaling to libdatadog
This will remove upscaling for `alloc-size` and `alloc-samples` profile
types from the library and make use of the upscaling capabilities in
`libdatadog`.
Doing this solves the rounding issues with the current implementation
due to float -> integer conversion, therefore allocation profiles will
be more correct and it offloads upscaling from the main thread to the
serialization step.