Sync profiling version with tracer; append git sha1 to non-release builds (#1992)
* Sync profiling version number with tracer

Also bump rust version in the Cargo.toml file. We had previously
already bumped the version in other places and just missed this one.

* Bump version to 0.86.3 on master

This is part of a change in our release process. Further changes will
be made to append a '+' symbol followed by the build hash.

* Append gitsha1 to version numbers
This doesn't happen on branches that begin with `ddtrace-`.