refactor: reverse split the ddprof_time threads
Also simplify the channel communications. There is a small risk here,
which is if the uploader thread has multiple profiles to upload, it
may now process them all before it joins the barrier, which was less
likely but possible before (priority was chosen by random chance
before, now it's message ordering).