fix: lifetime issue cached filename
The function part of the cache already had fixed this issue, but it
seems the filename portion was overlooked.

With this, there are no reasons anymore to have a Cow<str> instead of
a String for filename specifically, so I change that in the ZendFrame
struct as well.

perf: drop filename from the run time cache

Since it was now getting copied every time, and the logic to obtain
it is not complex, it doesn't justify the usage of the cache.

test: attempt to fix flaky test