refactor!: zai_string_view APIs and use them more
This adds new macros and functions for working with zai_string_view.

It changes the signatures of many functions taking pairs of pointers
and lengths to using zai_string_views instead, and the functions were
renamed because they had `literal` in their names which didn't apply
anymore. A small number of these already had versions that took views,
so the versions with literals were outright removed.