build(profiling): enable cross-language LTO
The profiler is mostly written in Rust and has a few small bits in C.
Most of these functions in C would be inlined if they were written in
the same language so the compilers could see the definitions.

Enter cross-language link-time optimization, or cross-language LTO.
This requires a compatible C toolchain and some LLVM tools.