Refactor zai_string_view APIs and use them more (#2203)
* refactor!: zai_string_view APIs and use them more

This adds new macros and functions for working with zai_string_view.

It changes the signatures of many functions taking pairs of pointers
and lengths to using zai_string_views instead, and the functions were
renamed because they had `literal` in their names which didn't apply
anymore. A small number of these already had versions that took views,
so the versions with literals were outright removed.

* refactor!: add zai_str_is_empty, remove zai_string_stuffed

* [ci skip] fix grammar in comment

* refactor: shorten the new zai_string_view APIs

This is in preparation for more renamings. Since these are new, the
diff has to be reviewed well anyway. For later renamings, it can be
reviewed quickly if there are basically just name changes.