Require `zai_str.ptr` to not be null (#2214)
* Require zai_str.ptr to not be null

It also moves the .ptr to be first. This matches Rust's layout, as well
as struct iovec on most platforms. It also helped find some C++ code
which was using direct initialization, as that requires fields to be
initialized in the same order as the source, and were thereby broken by
the re-ordering.

Also fix Tea include order so that we don't get warnings.

* Fix potential heap-buffer-overflow when zai is used without an observer