fix(profiling): do not leak module entry on purpose
This was done in the early days of the move from C to Rust. The goal
was to avoid unsafe code, and that was one way of working around
specific requirements. This is a different way, still mosty safe, but
will not leak. The goal is to soon run leak and address sanitizers.