fix(profiling): do not leak module entry on purpose (#2291)
This was done in the early days of the move from C to Rust. The goal
at the time was to avoid unsafe code, and that was one way of working
around specific requirements. This is a different way, still mostly
safe code, but will not leak.

Note that this wasn't a leak that mattered to anything except
tooling. The goal is to soon run an NTS+ASAN build for the profiler,
and this would trigger the leak sanitizer (implied by address
sanitizer on Linux).