Handle glob as input for rules
Sampling rules now use glob pattern matching. We used to have regexes, so it's a breaking change.

To work around this, we add a new DD_TRACE_SAMPLING_RULES_FORMAT parameter. This parameter allows `regex` and  `glob`. `regex` is the default for retrocompatibility reasons.

That said we will drop this parameter in 1.0.

Glob matching was already implemented for single span samplling so I just factorized some code.