Avoid resolving unresolved functions and classes
This only may pose problems with DDTrace\Trace attributes.
In theory having the hook installed twice should not pose problems, but adding a hook to an unresolved class is at the very least not a supported scenario. Checking for the classes being actually linked.

This only affects PHP 8.0 and PHP 8.1.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>