Update to use `DDOG_CHARSLICE_C` and new `DDOG_CHARSLICE_C_BARE` helpers
**What does this PR do?**

This PR updated dd-trace-php to use the helpers that were changed/
introduced in https://github.com/DataDog/libdatadog/pull/347 .

TL;DR: `DDOG_CHARSLICE_C` includes the cast to `(ddog_CharSlice)` to
all uses of `(ddog_CharSlice) DDOG_CHARSLICE_C` in the codebase were
converted to not have the cast; and `DDOG_CHARSLICE_C_BARE` does
not have the cast, so all uses of `DDOG_CHARSLICE_C` without the
cast in the codebase were converted to this one.

**Motivation:**

Update to match the latest libdatadog changes.

**Additional Notes:**

This is my first dd-trace-php contribution I think?

**How to test the change?**

I'm guessing CI should be able to validate this (?), not very
familiar with our testing setup.