refactor(profiling): run-time PHP_VERSION update
This uses the `php_version` function I added in PHP 8.3 to grab the
version of PHP at run-time when it's available. This still uses the
Reflection module version as a fallback for older versions.

If we fail to get the version at run-time, then we just fall back to
the compile-time version and emit a warning.