Ensure DDTRACE_G(active_stack) is always NULL when the root span is being pushed
I have zero idea _how_ this happened, given that I'm unable to reproduce it, but a core dump indicated that the memory it pointed to was from the previous request.
While this is not a very proper fix (as this should not be possible in the first place), it mitigates the effects.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>