Generate telmetry metrics in appsec helper
Introduces the interface metrics::TelemetrySubmitter and uses two
implementations to submit metrics 1) RequestMetricsSubmitter: generated
during a request, which are forwarded to the extension in that
requests's rshutdown message and
2) MetricsImpl: generated outside a request, inside the service. These
are forwarded to the extension in either client_init or rshutdown of any
client of that service.

This service is also used to submit the legacy meta and span metrics.