Avoid repeated processing of internal functions
Internal functions, when inherited, may (on PHP 8.1 with opcache at least) have pointers which are into heap memory which is local to a given process (e.g. attributes), not present on the target process. Thus accessing the attributes of a duplicated internal functions will crash there.

And on top of that, we anyway don't want to process these twice.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>