Add missing "in-preloading" check to `RSHUTDOWN`
`RSHUTDOWN` was missing the "in-preloading" check which would lead to a
situation where some profilers would collect data even though in
preloading. In allocation profiling this would lead to a situation where
if a customer has a custom memory handler (or uses `USE_ZEND_ALLOC=0`)
the allocation profiler would crash in
`allocation::alloc_prof_rshutdown()` when trying to reset the heap.