feat(prof): add 'tracing' features
This adds tracing features to the profiler. These are not meant to
be enabled for customer builds at this time. The goal is to use this
data to enable benchmarking and other useful tools during development.

Currently it's turned on by the RUST_LOG env var.

We may want to move some of our logging of the PHP lifecycle to use
this in the future.