Amend login failure event being reported by mistake (#3092)
It was reported that wodpress was reporting many login failure events when they should not be reported. After investigation it was spotted that whenever the login page was hit(even with a get), the login failure event was reported. This PR fixes that.

Additionally, user login events were not tested on Wordpress 6.1. This pr takes the opportunity to make those tests run