fix(prof): borrow error in request shutdown (#3247)
This affects PHP 8.0+ and web SAPIs. The string arena was borrowed to
check how much space the arena uses, and then it gets replaced.

It's hard to make a test case that triggers this edge, because you have
to accumulate over 2 MiB in the string cache. This is also why few
customers have hit it. In web SAPIs, the worker will often be rotated
out for various before this happens. I verified it manually locally
with a lower threshold to make sure it's fine now.