fix(prof): crash in ZEND_INIT_ARRAY
The engine currently does not save the opline in ZEND_INIT_ARRAY, so
under rare situations, that opline will be dangling, which can cause
crashes.

The fix works because by having a user opcode handler the engine will
save the opline before it calls the user opcode handler.

I do not have a reproducer for this one yet, but we've hit this issue
before so I'm fairly confident this needs to be done until there is
an upstream fix in PHP, then we can do version-specific mitigation.