fix(prof): crash in ZEND_INIT_ARRAY (#3255)
The engine currently does not save the opline in ZEND_INIT_ARRAY.
Usually if it's out-dated and the line-number is a bit stale, it's not
a big deal. But in some cases the opline will be dangling, and in those
cases we read invalid memory. This is a bug in the PHP VM that needs to
also be fixed upstream. Since PHP 8.1 and 8.2 are in security fixes
only, this will go into PHP 8.3 and 8.4, and we'll have to mitigate it
on <8.3.

The fix works because by having a user opcode handler the engine will
save the opline before it calls the user opcode handler. This has a
small performance penalty for this specific opcode.

I do not have a reproducer for this one yet, but we've hit this issue
before in other opcodes. I'm fairly confident this needs to be done
until there is an upstream fix in PHP, then we can restrict this
mitigation to version-specific ranges.