perf(prof): wake up the aggregator less
I noticed that SyncWaker::notify was taking about as much time as
walking the call stack. This seems odd so I'm attempting to use a
queue instead, and only _sometimes_ sending a message across the
channel telling the other thread to wake up and process the queue.

This was written from a mindset of NTS, needs some thought for ZTS
but I assume the problems are worse there.